backend/requirements.txt
fastapi
uvicorn[standard]
boto3
python-dotenv
python-multipart
SQLAlchemy
psycopg2-binary
python-jose[cryptography]
email_validator
requests
groq
cerebras-cloud-sdk
openai
python-docx
pypdf
numpy
pytz
sentence_transformers
faiss-cpu
